+++ examples/test/dynarray/test_dynarray_copy_badcount.pas
@@ -1,0 +1,7 @@
+program TestDynArrayCopyBadCount;
+var
+    a, b: array of integer;
+begin
+    SetLength(a, 3);
+    b := Copy(a, 0, 'x');
+end.