+++ examples/test/dynarray/test_dynarray_copy_badstart.pas
@@ -1,0 +1,7 @@
+program TestDynArrayCopyBadStart;
+var
+    a, b: array of integer;
+begin
+    SetLength(a, 3);
+    b := Copy(a, 'x');
+end.